[CBRD-24598] Add new test case for memory allocation. (#1358)
http://jira.cubrid.org/browse/CBRD-24598

This issue only verified the core dump problem.
But now the order of executing the 'execute q... ' query occurs an error caused by the executing order.
ex )
execute q using 'A'; /* success /
execute q using 1; / success */

execute q using 'A'; /* fail - ERROR: Cannot coerce a value of domain "character" to domain "double". */

This will be discussed and fixed in another issue.
Until fix the Actual result will be /* ERROR -181 : Cannot coerce value of domain "character" to domain "double". */.
It will be the answer sheet and after the issue is solved it must be revised as success result.