[CBRD-24046] Result cache for correlated scalar subquery (#1747)
http://jira.cubrid.org/browse/CBRD-24046

A correlated subquery is executed for each row of the main query to get a value. If there are many duplicated inputs, the same query is repeatedly executed, which may cause a decrease in performance. We eliminate the cause of the slowdown through the result cache.

This PR is a clone of #1737, and the previous PR was mistakenly merged, so it was reverted and this new PR was created. These cases have completed the code review process.