[CBRD-25361] Fixed the issue where trace information for CTEs and subqueries not provided. (#1748)
http://jira.cubrid.org/browse/CBRD-25361

There are currently issues with trace statistics. When tracing queries that use UNION, DIFFERENCE, INTERSECTION, and CTE together, the statistical information for the CTE is not output. Additionally, when using UNION-like SQL statements, the trace for the total cost is also not output. In this issue, we have made changes to enable the output of trace for the total cost of UNION-like SQL statements and to ensure that the statistical information for all subqueries bundled under the UNION statement is output.

This PR is a clone of #1738, and the previous PR was mistakenly merged, so it was reverted and this new PR was created. These cases have completed the code review process.