[CBRD-25447] Support parallel heap scan (#2184)
http://jira.cubrid.org/browse/CBRD-25447
http://jira.cubrid.org/browse/CBRD-25977
http://jira.cubrid.org/browse/CBRD-25997
http://jira.cubrid.org/browse/CBRD-26011
To improve the performance of heap scans, we plan to add a parallel heap scan feature to CUBRID.

Currently, parallel heap scan is only available for tables with more than 32 pages, so the impact of this change will be limited.

Test cases without an ORDER BY clause may cause errors when using parallel heap scan.