[CUBRIDQA-1137] modified the ordering of deallocate statements to come before any subsequent prepare statements (#2216)
The “1069 prepared statement limit exceeded” error is still occurring in the current JDBC compatibility test. In this test (_11_session_var_05.sql), eight prepared statements are created at once, surpassing the limit. To prevent this, we modified the script so that each prepared statement (e.g., st1) is created, executed, and deallocated before the next one (e.g., st2) is created.