[CBRD-25776] Modified and revised a test case for Inline CTE support (#2232)
http://jira.cubrid.org/browse/CBRD-25776
Due to the issue above, CTEs (WITH clauses) that are referenced only once (i.e., not referenced multiple times outside the WITH clause) are rewritten as inline views.

To ensure consistent verification with previous TC (test cases), we have added a version of the same query with the ‘Materialize’ hint to prevent the query from being rewritten.

For reference, since the standard SQL execution using JDBC (index_ls_hint.answer) is unable to output the query plan, the answer files are managed separately as .answer and .answer_cci.